+++ .github/scripts/setup-vars.js
