+++ .github/workflows/ci-dev.yaml
@@ -84,7 +84,7 @@ jobs:
         id: vars
         with:
           script: |
-            const { default: setupVars } = await import('${{ github.workspace }}/.github/scripts/vars-setup.js')
+            const { default: setupVars } = await import('${{ github.workspace }}/.github/scripts/setup-vars.js')
             const setup = JSON.parse(process.env.CI_SETUP);
             return await setupVars({core, setup})
       - uses: google-github-actions/get-secretmanager-secrets@e5bb06c2ca53b244f978d33348d18317a7f263ce # v2
